--- apps/person-search-service/src/main/java/no/nav/registre/testnorge/personsearchservice/controller/search/BostedsadresseSearch.java
@@ -1,13 +1,0 @@
-package no.nav.registre.testnorge.personsearchservice.controller.search;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class BostedsadresseSearch {
-    String kommunenummer;
-    String postnummer;
-}
